add: Auth User

+++ src/controllers/cards.ts
@@ -2,7 +2,9 @@ import { Request, Response, NextFunction } from 'express';
 import { ObjectId } from 'mongoose';
 import card from '../models/card';
 import { IRequest } from '../types/types';
-import { dataError, defaultError, notFoundError } from '../error/error';
+import {
+  forbiddenError, dataError, defaultError, notFoundError,
+} from '../error/error';
 
 interface ICardController {
   getCards(
@@ -40,6 +42,7 @@ class CardController implements ICardController {
   async getCards(req: IRequest, res: Response, next: NextFunction) {
     try {
       const cards = await card.find({});
+
       return res.send({ data: cards });
     } catch (error) {
       return next(defaultError('Ошибка сервера!'));
@@ -48,13 +51,21 @@ class CardController implements ICardController {
 
   async deleteCardById(req: IRequest, res: Response, next: NextFunction) {
     try {
+      const id = req.user?._id;
       const { cardId } = req.params;
+
+      const findCard = await card.findById(cardId);
+      if (id !== findCard?.owner.toString()) {
+        return next(forbiddenError('Нет прав для удаления!'));
+      }
+
       const deleteCard = await card.findByIdAndDelete(cardId);
       if (!deleteCard) {
         return next(notFoundError('Карточка с указанным _id не найдена!'));
       }
+
       return res.send({
-        message: 'Карточка удалена',
+        message: 'Карточка удалена!',
       });
     } catch (error) {
       if (error instanceof Error && error.name === 'CastError') {
@@ -62,6 +73,7 @@ class CardController implements ICardController {
           dataError('Передан некорректный id!'),
         );
       }
+
       return next(defaultError('Ошибка сервера!'));
     }
   }
@@ -70,7 +82,9 @@ class CardController implements ICardController {
     try {
       const id = req.user?._id;
       const { name, link } = req.body;
+
       const newCard = await card.create({ name, link, owner: id });
+
       return res.send(newCard);
     } catch (error) {
       if (error instanceof Error && error.name === 'ValidationError') {
@@ -78,6 +92,7 @@ class CardController implements ICardController {
           dataError('Переданы некорректные данные при создании карточки!'),
         );
       }
+
       return next(defaultError('Ошибка сервера!'));
     }
   }
@@ -86,6 +101,7 @@ class CardController implements ICardController {
     try {
       const id = req.user?._id;
       const { cardId } = req.params;
+
       const likeCard = await card.findByIdAndUpdate(
         cardId,
         { $addToSet: { likes: id } },
@@ -94,6 +110,7 @@ class CardController implements ICardController {
       if (!likeCard) {
         return next(notFoundError('Передан несуществующий _id карточки!'));
       }
+
       return res.send(likeCard);
     } catch (error) {
       if (error instanceof Error && error.name === 'CastError') {
@@ -101,6 +118,7 @@ class CardController implements ICardController {
           dataError('Переданы некорректные данные для постановки лайка!'),
         );
       }
+
       return next(defaultError('Ошибка сервера!'));
     }
   }
@@ -109,6 +127,7 @@ class CardController implements ICardController {
     try {
       const id = req.user?._id as ObjectId;
       const { cardId } = req.params;
+
       const dislikeCard = await card.findByIdAndUpdate(
         cardId,
         { $pull: { likes: id } },
@@ -117,6 +136,7 @@ class CardController implements ICardController {
       if (!dislikeCard) {
         return next(notFoundError('Передан несуществующий _id карточки!'));
       }
+
       return res.send(dislikeCard);
     } catch (error) {
       if (error instanceof Error && error.name === 'CastError') {
@@ -124,6 +144,7 @@ class CardController implements ICardController {
           dataError('Переданы некорректные данные для снятии лайка!'),
         );
       }
+
       return next(defaultError('Ошибка сервера!'));
     }
   }
